refactor: rename DayCard component
